FPGAProxyBlog/Edge-FPGAinCloud/fpgascoring/modules/fpgaproxy/requirements.txt
azureml-accel-models[cpu]